crypto_sentiment_agency/requirements.txt
agency-swarm
praw
nltk
spacy
requests
python-dotenv
pyyaml